# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Escolar",
        "description": "API para sistema de gerenciamento escolar multi-tenant. Permite cadastro de usuários, alunos, turmas, diários, eventos, entre outros. O contexto do tenant é definido pelo header 'x-tenant-id'.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "alexseyf66@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Servidor Local"
        },
        {
            "url": "https://elo-api-alexseyfs-projects.vercel.app",
            "description": "Servidor de Produção"
        }
    ],
...
    "paths": {
        "/login": {
            "post": {
                "tags": ["Autenticação"],
                "summary": "Login de usuário (Tenant)",
                "description": "Autentica um usuário de uma escola específica.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "senha": { "type": "string" }
                                },
                                "required": ["email", "senha"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login realizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } },
                    "400": { "description": "Dados inválidos" },
                    "401": { "description": "Credenciais inválidas" }
                }
            }
        },
...
        "/api/v1/platform/login": {
            "post": {
                "tags": ["Plataforma"],
                "summary": "Login de Administrador da Plataforma",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "senha": { "type": "string" }
                                },
                                "required": ["email", "senha"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login realizado" },
                    "401": { "description": "Não autorizado" }
                }
            }
        },
        "/api/v1/platform/schools": {
            "get": {
                "tags": ["Plataforma"],
                "summary": "Listar todas as escolas",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Lista de escolas" },
                    "403": { "description": "Acesso negado (Requer PLATFORM_ADMIN)" }
                }
            }
        },
        "/api/v1/platform/users": {
            "get": {
                "tags": ["Plataforma"],
                "summary": "Listar todos os usuários da plataforma",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Lista de usuários" },
                    "403": { "description": "Acesso negado" }
                }
            }
        },
...
        "/usuarios": {
            "get": {
                "tags": ["Usuarios"],
                "summary": "Listar usuários",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Lista de usuários" }
                }
            },
            "post": {
                "tags": ["Usuarios"],
                "summary": "Criar usuário",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/UsuarioInput" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Usuário criado" },
                    "400": { "description": "Erro de validação" },
                    "409": { "description": "Email já registrado" }
                }
            }
        },
        "/usuarios/usuario-logado": {
            "get": {
                "tags": ["Usuarios"],
                "summary": "Obter dados do usuário logado",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Dados do usuário" },
                    "404": { "description": "Usuário não encontrado" }
                }
            }
        },
        "/usuarios/{usuarioId}": {
            "get": {
                "tags": ["Usuarios"],
                "summary": "Obter usuário por ID",
                "parameters": [
                    { "name": "usuarioId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Dados do usuário" },
                    "404": { "description": "Usuário não encontrado" }
                }
            }
        },
...
        "/turmas": {
            "get": {
                "tags": ["Turmas"],
                "summary": "Listar turmas",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de turmas" } }
            },
            "post": {
                "tags": ["Turmas"],
                "summary": "Criar turma",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": { "application/json": { "schema": { "type": "object", "properties": { "nome": { "type": "string" } } } } }
                },
                "responses": { "201": { "description": "Turma criada" } }
            }
        },
        "/turmas/totalAlunosTurma": {
            "get": {
                "tags": ["Turmas"],
                "summary": "Total de alunos por turma",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Totais obtidos" } }
            }
        },
        "/turmas/{usuarioId}/professor": {
            "post": {
                "tags": ["Turmas"],
                "summary": "Adicionar professor à turma",
                "parameters": [
                    { "name": "usuarioId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": { "application/json": { "schema": { "type": "object", "properties": { "turmaId": { "type": "integer" } }, "required": ["turmaId"] } } }
                },
                "responses": { "201": { "description": "Professor vinculado" } }
            }
        },
...
        "/professores/{professorId}/turmas": {
            "get": {
                "tags": ["Professores"],
                "summary": "Turmas de um professor",
                "parameters": [
                    { "name": "professorId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de turmas" } }
            }
        },
...
        "/professores/{professorId}/turma/{turmaId}": {
            "delete": {
                "tags": ["Professores"],
                "summary": "Desvincular professor de turma",
                "parameters": [
                    { "name": "professorId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Desvinculado com sucesso" } }
            }
        },
...
        "/eventos/turma/{turmaId}": {
            "get": {
                "tags": ["Eventos"],
                "summary": "Listar eventos da turma",
                "parameters": [
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de eventos" } }
            }
        },
...
        "/objetivos/turma/{turmaId}": {
            "get": {
                "tags": ["Objetivos"],
                "summary": "Objetivos da turma",
                "parameters": [
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista" } }
            }
        },
...
    },
    "components": {
        "parameters": {
            "TenantId": {
                "name": "x-tenant-id",
                "in": "header",
                "description": "Slug ou ID da escola (Tenant)",
                "required": true,
                "schema": { "type": "string" }
            }
        },
        "schemas": {
            "UsuarioInput": {
                "type": "object",
                "properties": {
                    "nome": { "type": "string" },
                    "email": { "type": "string" },
                    "senha": { "type": "string" },
                    "telefone": { "type": "string" },
                    "roles": { "type": "array", "items": { "type": "string", "enum": ["ADMIN", "PROFESSOR", "RESPONSAVEL"] } },
                    "schoolId": { "type": "string", "description": "Obrigatório para PLATFORM_ADMIN criando usuário" }
                },
                "required": ["nome", "email", "senha", "telefone", "roles"]
            },
...
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "token": { "type": "string" },
                    "id": { "type": "integer" },
                    "nome": { "type": "string" },
                    "email": { "type": "string" },
                    "roles": { "type": "array", "items": { "type": "string" } },
                    "schoolSlug": { "type": "string" }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}